Part II & Part III: Integration test of canal. (#967)
* only integration

* only integration

* remove prev code

* remove prev code

* feat lint
