+++ docker/config/canal-test-config.toml
@@ -1,0 +1,5 @@
+enable-old-value = true
+
+[sink]
+protocol = "canal"
+
