+++ integration/integration.go
@@ -17,23 +17,29 @@ import (
 	"flag"
 	"github.com/pingcap/log"
 	"github.com/pingcap/ticdc/integration/framework"
+	canal2 "github.com/pingcap/ticdc/integration/framework/canal"
+	"github.com/pingcap/ticdc/integration/tests/avro"
+	"github.com/pingcap/ticdc/integration/tests/canal"
+
+	avro2 "github.com/pingcap/ticdc/integration/framework/avro"
+
 	"go.uber.org/zap/zapcore"
 )
 
-func main() {
-	dockerComposeFile := flag.String("docker-compose-file", "", "the path of the Docker-compose yml file")
+var dockerComposeFile = flag.String("docker-compose-file", "", "the path of the Docker-compose yml file")
 
+func testAvro() {
 	testCases := []framework.Task{
-		newSimpleCase(),
-		newDeleteCase(),
-		newManyTypesCase(),
-		newUnsignedCase(),
-		newCompositePKeyCase(),
-		newAlterCase(), // this case is slow, so put it last
+		avro.NewSimpleCase(),
+		avro.NewDeleteCase(),
+		avro.NewManyTypesCase(),
+		avro.NewUnsignedCase(),
+		avro.NewCompositePKeyCase(),
+		avro.NewAlterCase(), // this case is slow, so put it last
 	}
 
 	log.SetLevel(zapcore.DebugLevel)
-	env := framework.NewAvroKafkaDockerEnv(*dockerComposeFile)
+	env := avro2.NewKafkaDockerEnv(*dockerComposeFile)
 	env.Setup()
 
 	for i := range testCases {
@@ -44,4 +50,33 @@ func main() {
 	}
 
 	env.TearDown()
+}
+
+func testCanal() {
+	testCases := []framework.Task{
+		//canal.NewSimpleCase(),
+		//canal.NewDeleteCase(),
+		canal.NewManyTypesCase(),
+		canal.NewUnsignedCase(),
+		canal.NewCompositePKeyCase(),
+		canal.NewAlterCase(), // this case is slow, so put it last
+	}
+
+	log.SetLevel(zapcore.DebugLevel)
+	env := canal2.NewKafkaDockerEnv(*dockerComposeFile)
+	env.Setup()
+
+	for i := range testCases {
+		env.RunTest(testCases[i])
+		if i < len(testCases)-1 {
+			env.Reset()
+		}
+	}
+
+	env.TearDown()
+}
+
+func main() {
+	testAvro()
+	testCanal()
 }
